[HVZK-WHIR 3/6] split zk.rs into module folder and apply review cleanup
Split the 1644-line whir/src/sumcheck/zk.rs into six files under
whir/src/sumcheck/zk/: mod.rs (re-exports + protocol overview),
data.rs (transcript schema + MaskOracle alias), prover.rs, verifier.rs,
simulator.rs, and test_helpers.rs (shared fixtures across the three
test modules).

Refactor the witness-free simulator to take &ZkVerifier directly
instead of a pre-computed mu, eliminating the fork-clone-peek
choreography test drivers needed to extract alpha. Add a pub(crate)
ZkVerifier::sum that forwards to the inner verifier so the simulator
can derive mu without re-implementing the alpha-power loop.

Replace prop_simulator_f_subspace_invariant with prop_simulator_invariants
covering four checks per case: shape (round count, wire length,
ell_zk header, pow_witnesses, commits, gammas), positive stratification
(wire[i >= 2] in F-subspace), negative stratification (wire[0..2]
escapes F-subspace with overwhelming probability so the two tiers
have not collapsed), and verifier acceptance (cloned-challenger replay
of the simulated transcript).

Performance touch-ups: cross-field arithmetic via Algebra<F> avoids
EF::from(mu_tilde)-style materialisation, alpha and 2^i powers come
from Powers::collect_n with packed-field batching, sum_endpoints_init
uses mask[0].double() + mask[1..].sum() instead of mask[0] + sum,
and the LayoutStrategy is inlined at ZkVerifier::new rather than
hidden behind a helper.

Tighten doc comments and inline narration across all six files per
/doc rules: one sentence per line, ASCII formulas and flow blocks,
# Why mini-sections, no decorative banners.
